Moved utities to dreamer_utils and added comments on main file
