+++ examples/dreamer/dreamer_utils.py
@@ -40,6 +40,8 @@ LIBS = {
     "gym": GymEnv,
     "dm_control": DMControlEnv,
 }
+import torch
+from torch.cuda.amp import autocast
 
 
 def make_env_transforms(
@@ -267,6 +269,104 @@ def parallel_env_constructor(
         env = transformed_env_constructor(**kwargs)()
         return env
     return parallel_env
+
+
+def recover_pixels(pixels, stats):
+    return (
+        (255 * (pixels * stats["scale"] + stats["loc"]))
+        .clamp(min=0, max=255)
+        .to(torch.uint8)
+    )
+
+
+@torch.inference_mode()
+def call_record(
+    logger,
+    record,
+    collected_frames,
+    sampled_tensordict,
+    stats,
+    model_based_env,
+    actor_model,
+    cfg,
+):
+    td_record = record(None)
+    if td_record is not None and logger is not None:
+        for key, value in td_record.items():
+            if key in ["r_evaluation", "total_r_evaluation"]:
+                logger.log_scalar(
+                    key,
+                    value.detach().item(),
+                    step=collected_frames,
+                )
+    # Compute observation reco
+    if cfg.record_video and record._count % cfg.record_interval == 0:
+        world_model_td = sampled_tensordict
+
+        true_pixels = recover_pixels(world_model_td["next_pixels"], stats)
+
+        reco_pixels = recover_pixels(world_model_td["next_reco_pixels"], stats)
+        with autocast(dtype=torch.float16):
+            world_model_td = world_model_td.select("state", "belief", "reward")
+            world_model_td = model_based_env.rollout(
+                max_steps=true_pixels.shape[1],
+                policy=actor_model,
+                auto_reset=False,
+                tensordict=world_model_td[:, 0],
+            )
+        imagine_pxls = recover_pixels(
+            model_based_env.decode_obs(world_model_td)["next_reco_pixels"],
+            stats,
+        )
+
+        stacked_pixels = torch.cat([true_pixels, reco_pixels, imagine_pxls], dim=-1)
+        if logger is not None:
+            logger.log_video(
+                "pixels_rec_and_imag",
+                stacked_pixels.detach().cpu(),
+            )
+
+
+def grad_norm(optimizer: torch.optim.Optimizer):
+    sum_of_sq = 0.0
+    for pg in optimizer.param_groups:
+        for p in pg["params"]:
+            sum_of_sq += p.grad.pow(2).sum()
+    return sum_of_sq.sqrt().detach().item()
+
+
+def make_recorder_env(cfg, video_tag, stats, logger, create_env_fn):
+    recorder = transformed_env_constructor(
+        cfg,
+        video_tag=video_tag,
+        norm_obs_only=True,
+        stats=stats,
+        logger=logger,
+        use_env_creator=False,
+    )()
+
+    # remove video recorder from recorder to have matching state_dict keys
+    if cfg.record_video:
+        recorder_rm = TransformedEnv(recorder.base_env)
+        for transform in recorder.transform:
+            if not isinstance(transform, VideoRecorder):
+                recorder_rm.append_transform(transform)
+    else:
+        recorder_rm = recorder
+
+    if isinstance(create_env_fn, ParallelEnv):
+        recorder_rm.load_state_dict(create_env_fn.state_dict()["worker0"])
+        create_env_fn.close()
+    elif isinstance(create_env_fn, EnvCreator):
+        recorder_rm.load_state_dict(create_env_fn().state_dict())
+    else:
+        recorder_rm.load_state_dict(create_env_fn.state_dict())
+    # reset reward scaling
+    for t in recorder.transform:
+        if isinstance(t, RewardScaling):
+            t.scale.fill_(1.0)
+            t.loc.fill_(0.0)
+    return recorder
 
 
 @dataclass
