+++ torchrl/trainers/helpers/models.py
@@ -1436,6 +1436,7 @@ def make_dreamer(
 
 @dataclass
 class DreamerConfig:
+    batch_length: int = 50
     state_dim: int = 30
     rssm_hidden_dim: int = 200
     mlp_num_units: int = 400
@@ -1444,11 +1445,6 @@ class DreamerConfig:
     actor_value_lr: float = 8e-5
     imagination_horizon: int = 15
     model_device: str = ""
-    normalize_rewards_online: bool = False
-    # Computes the running statistics of the rewards and normalizes them before they are passed to the loss module.
-    normalize_rewards_online_scale: float = 1.0
-    # Final scale of the normalized rewards.
-    normalize_rewards_online_decay: float = 0.9999
     # Decay of the reward moving averaging
     exploration: str = "additive_gaussian"
     # One of "additive_gaussian", "ou_exploration" or ""
